+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - リソース作成の脆弱性対応
 - スケジュール検索の興行表示を最適化
 - rateLimitを追加
+- chevreの502リトライ対応
+- no-default-export: true
 
 ### Deprecated